Pipeline implemented and working, but doing nothing for the moment.
IMPORTANT :
Instead of getting the frames of the video while running the pipeline,
we get the frames at the beginning of the program and put them into a vector.
This is much easier to handle and the code is much cleaner.

Note : Some clean-up to do in the pipeline files.